requirements.txt
fastapi==0.125.0
uvicorn
langchain==1.1.3
langchain-community==0.4.1
langchain-core==1.1.3
langchain-text-splitters==1.0.0
langchain-anthropic==1.2.0
langchain-openai==1.1.2
openai==2.9.0
faiss-cpu==1.13.1
pydantic==2.10.3
pydantic_core==2.27.1
pydantic-settings==2.12.0
python-dotenv==1.1.0
pypdf==4.2.0
docx2txt==0.8
streamlit==1.45.1
requests==2.32.5